% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{aa_colors}
\alias{aa_colors}
\title{Amino acid color palette.}
\usage{
aa_colors()
}
\value{
A named character vector of hex colors.
}
\description{
Return a named character vector of colors for the 20 standard amino
acids plus fMet, Ile2, and SeC.
}
\examples{
aa_colors()
}
