% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{aa_props}
\alias{aa_props}
\title{Amino acid properties}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 5 columns.
}
\usage{
aa_props
}
\description{
Amino acid properties
}
\keyword{datasets}
