% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{available_organisms}
\alias{available_organisms}
\title{Available organisms with pre-computed global coordinates}
\usage{
available_organisms()
}
\value{
Character vector of organism names
}
\description{
Available organisms with pre-computed global coordinates
}
\examples{
available_organisms()
}
