% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{charging_colors}
\alias{charging_colors}
\title{Charging status color palette.}
\usage{
charging_colors()
}
\value{
A named character vector of hex colors.
}
\description{
Return a named character vector of colors for charged and uncharged
tRNA states.
}
\examples{
charging_colors()
}
