% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{classify_trna_type}
\alias{classify_trna_type}
\title{Classify tRNA as Type I or Type II}
\usage{
classify_trna_type(trna_ids)
}
\arguments{
\item{trna_ids}{Character vector of tRNA IDs (e.g., "tRNA-Leu-CAA-1-1")}
}
\value{
Character vector of "Type I" or "Type II"
}
\description{
Type II tRNAs have extended variable loops (9-24 extra nucleotides). These
include Leucine, Serine, Tyrosine, and Selenocysteine tRNAs. All other
tRNAs are Type I with standard 73-77 nucleotide length.
}
\examples{
# Type II tRNAs
classify_trna_type(c("tRNA-Leu-CAA-1-1", "tRNA-Ser-GCT-1-1"))

# Type I tRNAs
classify_trna_type(c("tRNA-Ala-GGC-1-1", "tRNA-Arg-ACG-1-1"))
}
