% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover.data.R
\name{clover.data}
\alias{clover.data}
\title{Extracts data and metadata from CLIMBER-X netcdf files}
\usage{
clover.data(dataDir, ncFileName, variableName)
}
\arguments{
\item{dataDir}{A string that gives the path where the netcdf file is located}

\item{ncFileName}{The name of the netcdf file (e.g. "atm.nc")}

\item{variableName}{variable name of interest (e.g. "amoc26N")}
}
\value{
A list with the data and metadata for a given variable
}
\description{
This function accesses the data and metadata stored in netcdf files for a given variable
}
