% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover.dev.dimensions.R
\name{clover.dev.dimensions}
\alias{clover.dev.dimensions}
\title{Creates table that summarizes file content}
\usage{
clover.dev.dimensions(dataDir)
}
\arguments{
\item{dataDir}{A string that gives the path where the netcdf file is located}

\item{ncFileName}{The name of the netcdf file}
}
\value{
Summary table of file content in markdown
}
\description{
This function finds all unique combinations of development across all CLIMBER-X outputs
}
