% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover.intFun.plot.lat.mon.time.R
\name{clover.intFun.plot.lat.mon.time}
\alias{clover.intFun.plot.lat.mon.time}
\title{Plot CLIMBER-X output}
\usage{
clover.intFun.plot.lat.mon.time(ncFileName, data, v, long_name, units, time)
}
\description{
Internal function used in clover.plot()
}
