% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover.plot.R
\name{clover.plot}
\alias{clover.plot}
\title{Plots content of CLIMBER-X netcdf files that contain more than just time series}
\usage{
clover.plot(dataDir, ncFileName)
}
\arguments{
\item{dataDir}{A string that gives the path where the netcdf file is located}

\item{ncFileName}{The name of the netcdf file (e.g. "atm.nc")}
}
\value{
Figues for each variable
}
\description{
This function plots all variables stored in a CLIMBER-X netcdf file that contains more than just time series.
Note that variables are stored using different combinations of dimensions.
}
