% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover.table.R
\name{clover.table}
\alias{clover.table}
\title{Creates table that summarizes file content}
\usage{
clover.table(dataDir, ncFileName)
}
\arguments{
\item{dataDir}{A string that gives the path where the netcdf file is located}

\item{ncFileName}{The name of the netcdf file}
}
\value{
Summary table of file content in markdown
}
\description{
This function creates a table in markdown format that summarizes the content of CLIMBER-X netcdf files that contain more than just time series (e.g. atm.nc)
}
