% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover.table.ts.R
\name{clover.table.ts}
\alias{clover.table.ts}
\title{Create a summary table for content of netcdf files with time series}
\usage{
clover.table.ts(dataDir, ncFileName)
}
\arguments{
\item{dataDir}{A string that gives the path where the netcdf file is located}

\item{ncFileName}{The name of the netcdf file}
}
\value{
An Table in markdown with all variables, units, and some basic statistics (min, max, mean, sd)
}
\description{
This function reads the content of CLIMBER-X netcdf files that contain time series (e.g. atm_ts.nc) and produces a table in markdown format that summarizes the file's content
}
