% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clover_example}
\alias{clover_example}
\title{Provide working directory for clover example files.}
\usage{
clover_example(path)
}
\arguments{
\item{path}{path to file}
}
\description{
Provide working directory for clover example files.
}
\examples{
clover_example("ecoli/config.yaml")

}
