% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cloverleaf.R
\docType{package}
\name{cloverleaf-package}
\alias{cloverleaf}
\alias{cloverleaf-package}
\title{cloverleaf: Plotting and analysis for tRNA-related data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselbrerth@cuanschutz.edu} (\href{https://orcid.org/0000-0002-6299-179X}{ORCID})

}
\keyword{internal}
