% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cloverleaf_example}
\alias{cloverleaf_example}
\title{Provide working directory for cloverleaf example files.}
\usage{
cloverleaf_example(path)
}
\arguments{
\item{path}{path to file}
}
\description{
Provide working directory for cloverleaf example files.
}
\examples{
cloverleaf_example("bcerror.tsv.gz")

}
