% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dna_to_rna_anticodon}
\alias{dna_to_rna_anticodon}
\title{Convert DNA anticodon to RNA in tRNA name strings}
\usage{
dna_to_rna_anticodon(x)
}
\arguments{
\item{x}{Character vector of tRNA name strings.}
}
\value{
Character vector with DNA anticodons converted to RNA.
}
\description{
Replaces T with U in the anticodon portion of tRNA identifiers,
e.g., \code{"tRNA-Glu-TTC-1-1"} becomes \code{"tRNA-Glu-UUC-1-1"}.
}
\examples{
dna_to_rna_anticodon("tRNA-Glu-TTC-1-1")
}
