% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{identity_organisms}
\alias{identity_organisms}
\title{List supported organisms for identity elements}
\usage{
identity_organisms()
}
\value{
A character vector of organism names.
}
\description{
Returns the names of organisms for which tRNA aminoacylation
identity element data is available.
}
\examples{
identity_organisms()
}
