% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kl_divergence}
\alias{kl_divergence}
\title{Kullback-Leibler divergence}
\usage{
kl_divergence(p, q)
}
\description{
Kullback-Leibler divergence
}
\examples{
# test for asymmetry
p <- c(0.9, 0.1, 0.0001, 0.0001)
q <- c(0.1, 0.9, 0.0001, 0.0001)

p_kl <- kl_divergence(p, q)
q_kl <- kl_divergence(q, p)

abs(p_kl - q_kl)

}
\seealso{
Other stats: 
\code{\link{symmetric_kl}()}
}
\concept{stats}
