% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modomics.R
\name{modomics_organisms}
\alias{modomics_organisms}
\title{List organisms with cached MODOMICS data}
\usage{
modomics_organisms()
}
\value{
A character vector of organism names.
}
\description{
Returns the names of organisms for which MODOMICS tRNA
modification data is bundled with the package. These organisms
can be used with \code{\link[=modomics_mods]{modomics_mods()}} without internet access.
}
\examples{
modomics_organisms()
}
