% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{propagate_error_diff}
\alias{propagate_error_diff}
\title{Propagate standard error for a difference.}
\usage{
propagate_error_diff(se_a, se_b)
}
\arguments{
\item{se_a}{Numeric vector of standard errors for the first value.}

\item{se_b}{Numeric vector of standard errors for the second value.}
}
\value{
Numeric vector of propagated standard errors.
}
\description{
Compute the standard error of \code{a - b} assuming independent errors.
}
\examples{
propagate_error_diff(1, 0.5)
}
