% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bcerror.R
\name{read_bcerror}
\alias{read_bcerror}
\title{Read base-calling error (bcerror) TSV files.}
\usage{
read_bcerror(bcerr_path)
}
\arguments{
\item{bcerr_path}{Path to bcerror file.}
}
\value{
A tibble
}
\description{
bcerror files are generated with \href{https://github.com/rnabioco/detectrms-rs}{detectrms}
}
\examples{
bcerr_path <- clover_example("yeast/grande.bcerr.tsv.gz")
read_bcerror(bcerr_path)

}
