% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover-se.R
\name{read_mod_annotations}
\alias{read_mod_annotations}
\title{Read modifications file}
\usage{
read_mod_annotations(mods)
}
\arguments{
\item{mods}{path to modifications file}
}
\value{
A tibble.
}
\description{
TSV file with 4 columns:
}
\details{
\enumerate{
\item \code{ref}
\item \code{pos}
\item \code{mod_full}
\item \code{mod1}
}
}
