% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-charging.R
\name{read_odds_ratios_multi}
\alias{read_odds_ratios_multi}
\title{Read odds ratio files for multiple samples.}
\usage{
read_odds_ratios_multi(paths)
}
\arguments{
\item{paths}{A named character vector of file paths. Names are used as
sample identifiers.}
}
\value{
A tibble with all samples combined and a \code{sample_id} column.
}
\description{
Read and combine odds ratio files from multiple samples into a single
tibble with a \code{sample_id} column.
}
\examples{
paths <- c(
  "wt-15-ctl-01" = clover_example(
    "ecoli/summary/tables/wt-15-ctl-01/wt-15-ctl-01.odds_ratios.tsv.gz"
  ),
  "wt-15-ctl-02" = clover_example(
    "ecoli/summary/tables/wt-15-ctl-02/wt-15-ctl-02.odds_ratios.tsv.gz"
  )
)
read_odds_ratios_multi(paths)
}
