% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-structure.R
\name{structure_organisms}
\alias{structure_organisms}
\title{List organisms with bundled tRNA structure SVGs}
\usage{
structure_organisms()
}
\value{
A character vector of organism names.
}
\description{
Returns the names of organisms for which tRNA cloverleaf structure
SVGs are bundled with the package. These organisms can be used
with \code{\link[=plot_tRNA_structure]{plot_tRNA_structure()}}.
}
\examples{
structure_organisms()
}
