% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{symmetric_kl}
\alias{symmetric_kl}
\title{Symmetric Kullback-Leibler divergence}
\usage{
symmetric_kl(p, q)
}
\description{
Symmetric Kullback-Leibler divergence
}
\examples{
p <- c(0.3, 0.2, 0.2, 0.3)
q <- c(0.25, 0.25, 0.25, 0.25)

symmetric_kl(p, q)

}
\seealso{
Other stats: 
\code{\link{kl_divergence}()}
}
\concept{stats}
