% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{trna_consensus_labels}
\alias{trna_consensus_labels}
\title{Labels for consensus tRNA secondary structure}
\format{
An object of class \code{character} of length 96.
}
\usage{
trna_consensus_labels
}
\description{
Labels for consensus tRNA secondary structure
}
\keyword{datasets}
