% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprinzl.R
\name{trna_regions}
\alias{trna_regions}
\title{tRNA structural regions.}
\usage{
trna_regions()
}
\value{
A named list of integer vectors.
}
\description{
Return a named list mapping canonical tRNA structural region names
to integer vectors of Sprinzl position numbers.
}
\examples{
trna_regions()
}
