% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover-package.R
\docType{package}
\name{clover-package}
\alias{clover}
\alias{clover-package}
\title{clover: Plotting and analysis for nanopore tRNA sequencing data}
\description{
Analyze direct tRNA sequencing data. Provides differential expression, modification, and tools for visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rnabioco.github.io/clover}
  \item \url{https://github.com/rnabioco/clover}
}

}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselberth@cuanschutz.edu} (\href{https://orcid.org/0000-0002-6299-179X}{ORCID})

Other contributors:
\itemize{
  \item RNA Bioscience Initiative (https://ror.org/03wmf1y16) [funder, copyright holder]
}

}
\keyword{internal}
