% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover.intFun.plot.latv.lev.month.time.R,
%   R/clover.intFun.plot.latv.zlev.mon.time.R
\name{clover.intFun.plot.latv.zlev.mon.time}
\alias{clover.intFun.plot.latv.zlev.mon.time}
\title{Plot CLIMBER-X output}
\usage{
clover.intFun.plot.latv.zlev.mon.time(
  ncFileName,
  data,
  v,
  long_name,
  units,
  time,
  dimensions,
  dim_values
)

clover.intFun.plot.latv.zlev.mon.time(
  ncFileName,
  data,
  v,
  long_name,
  units,
  time,
  dimensions,
  dim_values
)
}
\description{
Internal function used in clover.plot()

Internal function used in clover.plot()
}
