% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover-se.R
\name{CloverSE}
\alias{CloverSE}
\title{Create a SummarizedExperiment}
\usage{
CloverSE(counts, fasta, bcerror = NULL, ...)
}
\arguments{
\item{bam}{list of BAM files, names are sample names}

\item{pod5}{list of POD5 files, names are samples names}

\item{mods}{TSV file containing modification information}

\item{fa}{path to the FASTA reference used in alignment}
}
\description{
Create a SummarizedExperiment
}
\examples{
counts <- read_counts(clover_example("grande.counts.tsv.gz"))
bcerror <- read_bcerror(clover_example("grande.bcerror.tsv.gz"))
fasta <- read_fasta(clover_example("yeast/trna-ref.fa.gz"))

CloverSE(counts, fasta, bcerror)

}
