% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover-se.R
\name{CloverSE}
\alias{CloverSE}
\title{Create a SummarizedExperiment}
\usage{
CloverSE(counts, fasta, bcerror = NULL, bam = NULL, pod5 = NULL, mods = NULL)
}
\arguments{
\item{counts}{path to counts file}

\item{fasta}{path to the FASTA reference used in alignment}

\item{bcerror}{path to base-calling error file}

\item{bam}{list of BAM files, names are sample names}

\item{pod5}{list of POD5 files, names are samples names}

\item{mods}{TSV file containing modification information}
}
\description{
Create a SummarizedExperiment
}
