% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_clover}
\alias{create_clover}
\title{Create a RangedSummarizedExperiment object}
\usage{
create_clover(fa, bam, mods = NULL, pod5 = NULL)
}
\arguments{
\item{fa}{path to the FASTA reference used in alignment}

\item{bam}{list of BAM files, names are sample names}

\item{mods}{TSV file containing modification information}

\item{pod5}{list of POD5 files, names are samples names}
}
\description{
Create a RangedSummarizedExperiment object
}
