% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bcerror}
\alias{plot_bcerror}
\title{Plot base-calling error in a heatmap.}
\usage{
plot_bcerror(data, title_suffix, include_legend = TRUE)
}
\arguments{
\item{title_suffix}{string}

\item{include_legend}{Logical, include legend in the plot.}

\item{tbl}{tibble of bc-delta values}
}
\description{
Plot base-calling error in a heatmap.
}
