% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bcerror.R
\name{read_bcerror}
\alias{read_bcerror}
\title{Read base-calling error ("bcerror") TSV files.}
\usage{
read_bcerror(bcerr_path, raw = FALSE)
}
\arguments{
\item{bcerr_path}{Path to bcerror file.}

\item{raw}{Logical, return raw data.}
}
\value{
A tibble
}
\description{
bcerror files are generated with
\href{https://github.com/rnabioco/detectrms-rs}{detectrms}
}
\details{
\itemize{
\item \code{ref}: fasta reference name
\item \code{pos} : position in \code{ref}
\item \code{base}: reference base
\item \code{cov}: coverage at \code{pos}
\item \code{q_mean}, \code{q_median}, \code{q_std} : mean, median, and standard deviations of
quality scores
\item \code{error_rate} is the sum of \code{mis}, \code{ins}, and \code{del}.
\item \code{mis}, \code{ins}, \code{del}: frequency of mismatches, insertions, and deletions
\item \code{na}, \code{nc}, \code{ng}, \code{nt}: counts of A, C, G, T in the reads
}
}
\examples{
bcerr_path <- clover_example("yeast/grande.bcerr.tsv.gz")
read_bcerror(bcerr_path)

}
