% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bcerror.R
\name{read_counts}
\alias{read_counts}
\title{Read a counts TSV file.}
\usage{
read_counts(path)
}
\arguments{
\item{path}{Path to a counts TSV file (may be gzipped).}
}
\value{
A tibble.
}
\description{
Read a per-tRNA counts file produced by the tRNA sequencing pipeline.
}
\examples{
\dontrun{
read_counts("sample1.counts.tsv.gz")
}
}
