% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover-se.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA reference}
\usage{
read_fasta(fa)
}
\arguments{
\item{fa}{path to fasta file}
}
\value{
A \link[Biostrings:XStringSet-class]{Biostrings::DNAStringSet}.
}
\description{
Read FASTA reference
}
\examples{
fa <- clover_example("yeast/trna-ref.fa.gz")
read_fasta(fa)

}
