% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clover-se.R
\name{read_mod_annotations}
\alias{read_mod_annotations}
\title{Read modifications file}
\usage{
read_mod_annotations(mods)
}
\arguments{
\item{mods}{Path to a modifications TSV file.}
}
\value{
A tibble.
}
\description{
Read a TSV file with 4 columns: \code{ref}, \code{pos}, \code{mod_full}, \code{mod1}.
}
\examples{
path <- clover_example("ecoli/mod_annotations.tsv")
read_mod_annotations(path)
}
