% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-structure.R
\name{structure_trnas}
\alias{structure_trnas}
\title{List available tRNA structures for an organism}
\usage{
structure_trnas(organism)
}
\arguments{
\item{organism}{Character string specifying the organism name
(e.g., \code{"Escherichia coli"}). Use \code{\link[=structure_organisms]{structure_organisms()}} to
list available organisms.}
}
\value{
A character vector of tRNA names.
}
\description{
Returns the names of tRNAs for which cloverleaf structure SVGs
are bundled with the package for the given organism.
}
\examples{
\dontrun{
structure_trnas("Escherichia coli")
}
}
